+++ Assets/MixedRealityToolkit/Inspectors/PropertyDrawers/PhysicsLayerAttributeDrawer.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
+using Microsoft.MixedReality.Toolkit.Editor;
 using System.Collections.Generic;
 using UnityEditor;
 using UnityEngine;
@@ -18,9 +19,9 @@ namespace Microsoft.MixedReality.Toolkit.Physics.Editor
             var guiContents = new List<GUIContent>();
             var layerIds = new List<int>();
 
-            for (int i = 0; i < LayerExtensions.TagManagerLayers.arraySize; i++)
+            for (int i = 0; i < EditorLayerExtensions.TagManagerLayers.arraySize; i++)
             {
-                var layer = LayerExtensions.TagManagerLayers.GetArrayElementAtIndex(i);
+                var layer = EditorLayerExtensions.TagManagerLayers.GetArrayElementAtIndex(i);
 
                 if (!string.IsNullOrWhiteSpace(layer.stringValue))
                 {
